+++ packages/db/src/index.ts
@@ -1,5 +1,17 @@
 export { createDb, type AppDatabase } from "./client";
+export {
+  isSampleDataModeEnabled,
+  resolveDatabasePath,
+} from "./database-path";
 export { newId, nowIso } from "./id";
+export {
+  SAMPLE_IDS,
+  clearSampleData,
+  isSampleDataSeeded,
+  seedSampleData,
+  type ClearSampleDataResult,
+  type SeedSampleDataResult,
+} from "./sample-data";
 export * from "./schema/index";
 export * from "./repositories/portfolios";
 export * from "./repositories/instruments";